+++ app/src/androidTest/java/com/saloavalos/android_app_samir/ExampleInstrumentedTest.java
@@ -1,26 +1,24 @@
-package com.example.proyectosamir;
+package com.saloavalos.android_app_samir;
 
 import android.content.Context;
+
 import androidx.test.platform.app.InstrumentationRegistry;
-import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
 /**
  * Instrumented test, which will execute on an Android device.
  *
  * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
  */
-@RunWith(AndroidJUnit4.class)
 public class ExampleInstrumentedTest {
     @Test
     public void useAppContext() {
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
 
-        assertEquals("com.example.proyectosamir", appContext.getPackageName());
+        assertEquals("com.saloavalos.android_app_samir", appContext.getPackageName());
     }
 }
