+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.proyectosamir">
+    package="com.saloavalos.android_app_samir">
 
     <uses-feature
         android:name="android.hardware.camera"
@@ -17,10 +17,10 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".activities.SignInActivity"></activity>
-        <activity android:name=".activities.SignUpActivity" />
-        <activity android:name=".activities.SecondActivity" />
-        <activity android:name=".activities.MainActivity">
+        <activity android:name="com.saloavalos.android_app_samir.activities.SignInActivity"></activity>
+        <activity android:name="com.saloavalos.android_app_samir.activities.SignUpActivity" />
+        <activity android:name="com.saloavalos.android_app_samir.activities.SecondActivity" />
+        <activity android:name="com.saloavalos.android_app_samir.activities.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -31,13 +31,16 @@
         <!-- android:name="androidx.core.content.FileProvider" -->
         <provider
             android:name="androidx.core.content.FileProvider"
-            android:authorities="com.example.android.fileprovider"
+            android:authorities="com.saloavalos.android.fileprovider"
             android:exported="false"
             android:grantUriPermissions="true">
             <meta-data
                 android:name="android.support.FILE_PROVIDER_PATHS"
                 android:resource="@xml/file_paths" />
         </provider>
+        <meta-data
+            android:name="preloaded_fonts"
+            android:resource="@array/preloaded_fonts" />
     </application>
 
 </manifest>
