+++ app/src/main/java/com/saloavalos/android_app_samir/activities/MainActivity.java
@@ -1,4 +1,4 @@
-package com.example.proyectosamir.activities;
+package com.saloavalos.android_app_samir.activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -6,10 +6,9 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
-import android.widget.EditText;
 import android.widget.TextView;
 
-import com.example.proyectosamir.R;
+import com.saloavalos.android_app_samir.R;
 
 public class MainActivity extends AppCompatActivity {
