+++ app/src/main/java/com/saloavalos/android_app_samir/activities/SecondActivity.java
@@ -1,4 +1,4 @@
-package com.example.proyectosamir.activities;
+package com.saloavalos.android_app_samir.activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.app.ActivityCompat;
@@ -8,7 +8,6 @@ import androidx.core.content.FileProvider;
 import android.Manifest;
 import android.content.Intent;
 import android.content.pm.PackageManager;
-import android.graphics.Bitmap;
 import android.net.Uri;
 import android.os.Bundle;
 import android.os.Environment;
@@ -17,7 +16,7 @@ import android.util.Log;
 import android.view.View;
 import android.widget.ImageView;
 
-import com.example.proyectosamir.R;
+import com.saloavalos.android_app_samir.R;
 
 import java.io.File;
 import java.io.IOException;
