+++ app/src/main/res/layout/activity_main.xml
@@ -2,17 +2,16 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/tvinfo"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/color_bg_main"
     tools:context=".activities.MainActivity">
 
-    <androidx.appcompat.widget.AppCompatImageView
+    <ImageView
         android:id="@+id/img_main_banner"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="42dp"
+        android:layout_width="300dp"
+        android:layout_height="300dp"
+        android:layout_marginTop="28dp"
         android:src="@drawable/main_banner_img"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
@@ -46,12 +45,12 @@
 
     <Button
         android:id="@+id/btn_main_signup"
+        style="@style/btn_white"
         android:layout_width="match_parent"
         android:layout_height="55dp"
         android:layout_marginStart="16dp"
         android:layout_marginTop="42dp"
         android:layout_marginEnd="16dp"
-        android:background="@drawable/btn_manager_white"
         android:fontFamily="@font/open_sans_semibold"
         android:text="Registrarse"
         android:textColor="@color/color_black_general"
@@ -68,7 +67,7 @@
         android:fontFamily="@font/open_sans"
         android:text="¿tienes una cuenta?"
         android:textColor="@color/color_white"
-        android:textSize="21sp"
+        android:textSize="20sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintStart_toStartOf="parent" />
 
@@ -81,7 +80,7 @@
         android:fontFamily="@font/open_sans_semibold"
         android:text="INICIA SESIÓN"
         android:textColor="@color/color_white"
-        android:textSize="21sp"
+        android:textSize="20sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent" />
